[Fix] Compile fixes for libopenmpt and DJGPP.
git-svn-id: https://source.openmpt.org/svn/openmpt/trunk/OpenMPT@16432 56274372-70c3-4bfc-bfc3-4c3a0b034d27

+++ soundlib/ModInstrument.cpp
@@ -285,7 +285,7 @@ void ModInstrument::Sanitize(MODTYPE modType)
 	MPT_UNREFERENCED_PARAMETER(modType);
 	const uint8 range = ENVELOPE_MAX;
 #else
-	const uint8 range = modType == MOD_TYPE_AMS ? uint8_max : ENVELOPE_MAX;
+	const uint8 range = modType == MOD_TYPE_AMS ? uint8_max : uint8(ENVELOPE_MAX);
 #endif
 	VolEnv.Sanitize();
 	PanEnv.Sanitize();
